--- ayush_saxena_ece_8893_spring_26/Ayush_Saxena_lab_2_ECE_8893_spring_26/dataset_hls.tcl
@@ -1,27 +1,0 @@
-open_project project_1_optimized
-
-# set top function of the HLS design
-set_top compute_attention_HLS
-
-# add source file
-add_files top.cpp
-
-# add testbench
-add_files -tb host.cpp
-
-# add data file
-add_files -tb Q_tensor.bin
-add_files -tb V_tensor.bin
-add_files -tb K_tensor.bin
-add_files -tb Output_tensor.bin
-
-open_solution "solution1"
-
-# FPGA part and clock configuration
-set_part {xczu3eg-sbva484-1-e}
-
-# default frequency is 100 MHz
-#create_clock -period 4 -name default
-
-# C synthesis for HLS design, generating RTL
-csynth_design
