--- ayush_saxena_ece_8893_spring_26/Ayush_Saxena_lab_2_ECE_8893_spring_26/dcl.h
@@ -1,19 +1,0 @@
-
-#include <stdio.h>
-#include <iostream>
-#include <fstream>
-#include <cstdlib>
-#include <ap_fixed.h>
-#include <hls_math.h>
-#include <stdlib.h>
-#include <cstdint>
-
-typedef ap_fixed<16, 5> fixed_t;
-
-// Define tensor dimensions
-#define B   4      // Batch size
-#define N   100     // Sequence length
-#define dk  128     // Key/Query dimension
-#define dv  128     // Value dimension
-
-void compute_attention_HLS(fixed_t Q[B][N][dk], fixed_t K[B][N][dk], fixed_t V[B][N][dv], fixed_t Output[B][N][dv]);
