--- ayush_saxena_ece_8893_spring_26/Ayush_Saxena_lab_3_ECE_8893_spring_26/dataset_hls.tcl
@@ -1,26 +1,0 @@
-open_project project_1
-
-# set top function of the HLS design
-set_top sparse_matrix_multiply_HLS
-
-# add source file
-add_files top.cpp
-
-# add testbench
-add_files -tb host.cpp
-
-# add data file
-add_files -tb A_matrix_csr_sparsity_0.10.bin
-add_files -tb B_matrix_csc_sparsity_0.10.bin
-add_files -tb C_matrix_result_sparsity_0.10.bin
-
-open_solution "solution1"
-
-# FPGA part and clock configuration
-set_part {xczu3eg-sbva484-1-e}
-
-# default frequency is 100 MHz
-#create_clock -period 2 -name default
-
-# C synthesis for HLS design, generating RTL
-csynth_design
