+++ docs/source/built_in_datasets.md
@@ -104,3 +104,11 @@ HLSFactory provides several built-in datasets of HLS design sources.
 **Design Type:** Concrete
 
 **Supported Vendors:** Xilinx
+
+## Ayush Saxena Spring 26 ECE_88993
+
+**Number of Designs:** 2
+
+**Design Type:** Concrete
+
+**Supported Vendors:** Xilinx
