+++ hlsfactory/hls_dataset_sources/ayush_saxena_ece_8893_spring_26/Ayush_Saxena_lab_3_ECE_8893_spring_26/dcl.h
@@ -1,0 +1,22 @@
+
+#include <stdio.h>
+#include <iostream>
+#include <fstream>
+#include <cstdlib>
+#include <ap_fixed.h>
+#include <hls_math.h>
+#include <stdlib.h>
+#include <cstdint>
+
+typedef ap_fixed<16, 5> data_t;
+
+#define N 64  // Rows of A and C
+#define M 64  // Columns of A and Rows of B
+#define K 64  // Columns of B and C
+#define SPARSITY 0.5  // Percentage of non-zero elements
+
+
+void sparse_matrix_multiply_HLS(data_t values_A[N * M], int column_indices_A[N * M], int row_ptr_A[N + 1], 
+                             data_t values_B[M * K], int row_indices_B[M * K], int col_ptr_B[M + 1], data_t C[N][K]);
+
+
